# web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "FormFieldInput": {
        "additionalProperties": false,
        "properties": {
          "helpText": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "placeholder": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "short_text",
              "long_text",
              "single_select",
              "multi_select",
              "date",
              "yes_no",
              "consent_checkbox"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type",
          "required"
        ],
        "title": "FormFieldInput",
        "type": "object"
      },
      "FormTemplateInput": {
        "additionalProperties": false,
        "properties": {
          "attachedServiceIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Attachedserviceids",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FormFieldInput"
            },
            "maxItems": 100,
            "title": "Fields",
            "type": "array"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "requiredBeforePayment": {
            "title": "Requiredbeforepayment",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status",
          "fields",
          "requiredBeforePayment"
        ],
        "title": "FormTemplateInput",
        "type": "object"
      },
      "FormTemplateListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FormTemplateResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "FormTemplateListResponse",
        "type": "object"
      },
      "FormTemplatePatch": {
        "additionalProperties": false,
        "properties": {
          "attachedServiceIds": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachedserviceids"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FormFieldInput"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "requiredBeforePayment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiredbeforepayment"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "inactive"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "FormTemplatePatch",
        "type": "object"
      },
      "FormTemplateResponse": {
        "properties": {
          "attachedServiceIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Attachedserviceids",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FormFieldInput"
            },
            "title": "Fields",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requiredBeforePayment": {
            "title": "Requiredbeforepayment",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "fields",
          "attachedServiceIds",
          "requiredBeforePayment",
          "createdAt"
        ],
        "title": "FormTemplateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicFormListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicFormResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PublicFormListResponse",
        "type": "object"
      },
      "PublicFormResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FormFieldInput"
            },
            "title": "Fields",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requiredBeforePayment": {
            "title": "Requiredbeforepayment",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "fields",
          "requiredBeforePayment"
        ],
        "title": "PublicFormResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Slotera API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/forms": {
      "get": {
        "operationId": "listForms",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Forms",
        "tags": [
          "forms"
        ]
      },
      "post": {
        "operationId": "createForm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormTemplateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Form",
        "tags": [
          "forms"
        ]
      }
    },
    "/forms/{form_id}": {
      "delete": {
        "operationId": "deleteForm",
        "parameters": [
          {
            "in": "path",
            "name": "form_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Form Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Form",
        "tags": [
          "forms"
        ]
      },
      "get": {
        "operationId": "getForm",
        "parameters": [
          {
            "in": "path",
            "name": "form_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Form Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Form",
        "tags": [
          "forms"
        ]
      },
      "patch": {
        "operationId": "updateForm",
        "parameters": [
          {
            "in": "path",
            "name": "form_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Form Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormTemplatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Form",
        "tags": [
          "forms"
        ]
      }
    },
...
    "/public/workspaces/{workspace_slug}/services/{service_id}/forms": {
      "get": {
        "operationId": "listPublicServiceForms",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_slug",
            "required": true,
            "schema": {
              "title": "Workspace Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicFormListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Service Forms",
        "tags": [
          "public-booking"
        ]
      }
    },
...
  }
}